multi_language_invoice_extractor/requirements.txt
streamlit
PyPDF2
google.generativeai
python-dotenv
langchain
chromadb